% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred_2norm.R
\name{post_pred_2norm}
\alias{post_pred_2norm}
\title{A function that returns a list of two vectors containing posterior predictive parameters for a normal distribution with unknown variance and mean given a list of several vectors}
\usage{
post_pred_2norm(data_list, mu, nu, b, alpha)
}
\arguments{
\item{data_list}{list of numerical vectors}

\item{mu}{the mean of the prior distribution for the mean}

\item{nu}{the prior parameter that adjusts the conjugate variance of the mean}

\item{b}{the prior beta parameter that adjusts the conjugate variance of the likelihood distribution variance}

\item{alpha}{the prior alpha parameter that adjusts the conjugate variance of the likelihood distribution variance}
}
\value{
returns the posterior predictive parameters given the data and the priors
}
\description{
A function that returns a list of two vectors containing posterior predictive parameters for a normal distribution with unknown variance and mean given a list of several vectors
}
\examples{
post_pred_2norm(data_list, mu, nu, b, alpha)
}
